[CSV-1]
package 명변경
dimes 값 사용하도록 수정
Thumb drawable 에 그림자 추가
미사용 코드 제거
