+++ app/src/androidTest/java/com/study/switchtest/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.shinhan.switchtest
+package com.study.switchtest
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
