+++ app/src/main/res/drawable/thumb_circle_shadow.xml
@@ -1,0 +1,60 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:bottom="0dp"
+        android:left="0dp"
+        android:right="0dp">
+        <shape>
+            <solid android:color="#01000000" />
+            <corners android:radius="80dp" />
+        </shape>
+    </item>
+    <item
+        android:bottom="1dp"
+        android:left="1dp"
+        android:right="1dp">
+        <shape>
+            <solid android:color="#04000000" />
+            <corners android:radius="80dp" />
+        </shape>
+    </item>
+    <item
+        android:bottom="2dp"
+        android:left="2dp"
+        android:right="2dp">
+        <shape>
+            <solid android:color="#06000000" />
+            <corners android:radius="80dp" />
+        </shape>
+    </item>
+    <item
+        android:bottom="3dp"
+        android:left="3dp"
+        android:right="3dp">
+        <shape>
+            <solid android:color="#08000000" />
+            <corners android:radius="80dp" />
+        </shape>
+    </item>
+    <item
+        android:bottom="4dp"
+        android:left="4dp"
+        android:right="4dp">
+        <shape>
+            <solid android:color="#09000000" />
+            <corners android:radius="80dp" />
+        </shape>
+    </item>
+    <item
+        android:bottom="5dp"
+        android:left="5dp"
+        android:right="5dp">
+        <shape android:shape="rectangle">
+            <size
+                android:width="24dp"
+                android:height="24dp" />
+            <corners android:radius="80dp" />
+            <solid android:color="#fff" />
+        </shape>
+    </item>
+</layer-list>
