+++ app/src/main/res/layout/activity_main.xml
@@ -6,26 +6,7 @@
     android:layout_height="match_parent"
     tools:context=".MainActivity">
 
-    <!--    <androidx.appcompat.widget.SwitchCompat-->
-    <!--        android:id="@+id/switcher"-->
-    <!--        android:layout_width="64dp"-->
-    <!--        android:layout_height="24dp"-->
-    <!--        android:showText="true"-->
-    <!--        android:thumb="@drawable/thumb_circle"-->
-    <!--        android:gravity="center"-->
-    <!--        android:minWidth="64dp"-->
-    <!--        android:maxWidth="64dp"-->
-    <!--        android:minHeight="24dp"-->
-    <!--        android:maxHeight="24dp"-->
-    <!--        android:padding="-1dp"-->
-    <!--        app:switchMinWidth="24dp"-->
-    <!--        app:layout_constraintBottom_toBottomOf="parent"-->
-    <!--        app:layout_constraintLeft_toLeftOf="parent"-->
-    <!--        app:layout_constraintRight_toRightOf="parent"-->
-    <!--        app:layout_constraintTop_toTopOf="parent"-->
-    <!--        app:track="@drawable/switch_on_off" />-->
-
-    <com.shinhan.switchtest.SwitchView
+    <com.study.switchtest.SwitchView
         android:id="@+id/switchViewBtn"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
@@ -33,6 +14,5 @@
         app:layout_constraintLeft_toLeftOf="parent"
         app:layout_constraintRight_toRightOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
-
 
 </androidx.constraintlayout.widget.ConstraintLayout>
